+++ src/browserbehavior/DefaultBrowserBehavior.ts
@@ -17,6 +17,8 @@ export default class DefaultBrowserBehavior implements BrowserBehavior {
     safari: 12,
     opera: 66,
     samsung: 12,
+    crios: 86,
+    fxios: 23,
   };
 
   private browserName: { [id: string]: string } = {
@@ -28,6 +30,8 @@ export default class DefaultBrowserBehavior implements BrowserBehavior {
     safari: 'Safari',
     opera: 'Opera',
     samsung: 'Samsung Internet',
+    crios: 'Chrome iOS',
+    fxios: 'Firefox iOS',
   };
 
   private chromeLike: string[] = [
@@ -37,6 +41,9 @@ export default class DefaultBrowserBehavior implements BrowserBehavior {
     'opera',
     'samsung',
   ];
+
+  private webkitBrowsers: string[] = ['crios', 'fxios', 'safari', 'ios'];
+
   private enableUnifiedPlanForChromiumBasedBrowsers: boolean;
 
   constructor({
@@ -66,16 +73,26 @@ export default class DefaultBrowserBehavior implements BrowserBehavior {
     return false;
   }
 
+  hasWebKitWebRTC(): boolean {
+    for (const browser of this.webkitBrowsers) {
+      if (browser === this.browser.name) {
+        return true;
+      }
+    }
+    return false;
+  }
+
   hasFirefoxWebRTC(): boolean {
     return this.isFirefox();
   }
 
   supportsCanvasCapturedStreamPlayback(): boolean {
-    return !this.isIOSSafari();
+    return !this.isIOSSafari() && !this.isIOSChrome() && !this.isIOSFirefox();
   }
 
   requiresUnifiedPlan(): boolean {
-    let shouldEnable = (this.isSafari() && this.isUnifiedPlanSupported()) || this.isFirefox();
+    let shouldEnable =
+      this.isFirefox() || (this.hasWebKitWebRTC() && this.isUnifiedPlanSupported());
     if (this.enableUnifiedPlanForChromiumBasedBrowsers) {
       shouldEnable = shouldEnable || this.hasChromiumWebRTC();
     }
@@ -90,7 +107,7 @@ export default class DefaultBrowserBehavior implements BrowserBehavior {
   }
 
   requiresCheckForSdpConnectionAttributes(): boolean {
-    return !this.isIOSSafari();
+    return !this.isIOSSafari() && !this.isIOSChrome() && !this.isIOSFirefox();
   }
 
   requiresIceCandidateGatheringTimeoutWorkaround(): boolean {
@@ -98,7 +115,7 @@ export default class DefaultBrowserBehavior implements BrowserBehavior {
   }
 
   requiresUnifiedPlanMunging(): boolean {
-    let shouldRequire = this.isSafari() && this.isUnifiedPlanSupported();
+    let shouldRequire = this.hasWebKitWebRTC() && this.isUnifiedPlanSupported();
     if (this.enableUnifiedPlanForChromiumBasedBrowsers) {
       shouldRequire = shouldRequire || this.hasChromiumWebRTC();
     }
@@ -140,7 +157,9 @@ export default class DefaultBrowserBehavior implements BrowserBehavior {
     return this.isChrome() || this.isEdge();
   }
 
+  // TODO: Deprecated, needs to be removed
   screenShareUnsupported(): boolean {
+    console.warn('This function is no longer supported.');
     if (this.isSafari()) {
       return true;
     }
@@ -206,6 +225,14 @@ export default class DefaultBrowserBehavior implements BrowserBehavior {
 
   private isFirefox(): boolean {
     return this.browser.name === 'firefox';
+  }
+
+  private isIOSFirefox(): boolean {
+    return this.browser.name === 'fxios';
+  }
+
+  private isIOSChrome(): boolean {
+    return this.browser.name === 'crios';
   }
 
   private isChrome(): boolean {